feat/project structure (#1)
* feat: Added the base project layout and structure

This will for now serve as the base layout and
of the project.  There might likely still be uses for
a service layer.  We can adapt if need be

* feat: Added the base project layout and structure

This will for now serve as the base layout and
of the project.  There might likely still be uses for
a service layer.  We can adapt if need be